requirements.txt
fastapi>=0.119.1
uvicorn==0.38.0
starlette==0.49.1
pydantic==2.12.3
python-dotenv==1.1.1
httpx==0.28.1
requests==2.32.4
cryptography==44.0.1
PyJWT==2.10.1
bcrypt==5.0.0
pyyaml==6.0.3
click==8.1.7
numpy==2.3.4
psutil==7.1.1
structlog==25.4.0
orjson==3.11.3
passlib==1.7.4
redis==5.0.1
starlette-csrf==2.0.0
psycopg2-binary==2.9.9
slowapi==0.1.9
networkx==3.2.1
aiofiles==25.1.0
aiohttp==3.13.1
asyncio-mqtt==0.16.2
python-multipart==0.0.18
python-dateutil==2.9.0.post0
pytz==2025.2
sse-starlette==2.1.0
urllib3==2.6.0
liboqs-python==0.14.1
web3==7.14.0
prometheus-client==0.19.0
numba==0.63.1
shap==0.50.0
aioipfs==0.5.1
hnswlib==0.8.0
sentence-transformers==5.2.0
torch-geometric==2.7.0
flwr==1.25.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-redis==0.42b0
